Autodetect spidermonkey version in ./configure
I have struggled with this point when building
couchDB, because it's written nowhere in the documentation
that you need to configure the version properly.

It seems that other people had the problem as well:
https://couchdb.slack.com/archives/C49LEE7NW/p1632249155263400

The updated configure script automatically finds
out the installed library version. It also introduces
a dependency on `sed`, `ldconfig` and `head`,
and remove the default to version 1.8.5 which is old
and not in most distributions anymore.

The argument was also removed from the CI scripts
because it does not exist anymore.